bin: Add 'assemble' script to autogenerate Components from SVGs
We now have some codegen going! When the 'assemble' script is run, all
files in the /assets directory are iterated and the corresponding React
Components are generated for them. We currently place these in
/bin/icons, but when this feature is mature we will place them straight
into /src/icons.

In order to work properly, all weights for an icon should be grouped in
a directory with the icon name in kebab-case, and should use the
fully-qualified weight suffix, except in the case of 'regular' icons,
which should have no suffix. E.G.:

/assets
  /backspace
    backspace.svg
    backspace-bold.svg
    backsapce-duotone.svg
    backsapce-fill.svg
    backspace-light.svg
    backspace-thin.svg

To note: we still need to address the missing 'stroke' and strokeWidth'
attributes on <g> elements.